+++ recovery/recovery_test/log_recovery_test.go
@@ -153,8 +153,10 @@ func TestLogSererializeAndDeserialize(t *testing.T) {
 */
 
 func TestRedo(t *testing.T) {
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	samehada_instance := samehada.NewSamehadaInstanceForTesting()
 
@@ -255,9 +257,10 @@ func TestRedo(t *testing.T) {
 }
 
 func TestUndo(t *testing.T) {
-	os.Stdout.Sync()
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	samehada_instance := samehada.NewSamehadaInstanceForTesting()
 
@@ -398,8 +401,10 @@ func TestUndo(t *testing.T) {
 }
 
 func TestCheckpoint(t *testing.T) {
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 	samehada_instance := samehada.NewSamehadaInstanceForTesting()
 
 	testingpkg.AssertFalse(t, common.EnableLogging, "")
